[AMP-140 Fix] Correct FP constants for OpenCL
Floating point constants where not in the __constant address space, causing
kernel compilation to fail in some cases. The reinterpretation from bit pattern
in an unsigned int to floating point also needed to be marked as a pointer to
__constant memory. Neither of these changes are valid in standard C so they are
guarded in the outputted code by #ifdef __OPENCL_VERSION_ which the opencl
compiler in all stacks should define.