[AMP-150 Fix] member internal() missing restrict
The array member function internal() which returned the gmac pointer and is
used in constructing an array_view from an array was missing its
restrict(amp,cpu) specifier, so it was not included when compiling in a kernel
context. This triggered a crash in the promote pass and would have led to
undefined globals anyway.