[Refactor] Only synchronize when last view deleted.
Acording to the spec array_view only has an implicit synchronize when the last
copy falls out of scope. Before we where synchronizing whenever an array_view
was destructed, causing major performance issues in repeated parallel_for_each
calls that copied an array_view by value. The gmac cache is a std::shared_ptr,
so we use its use count to discover if we are the last copy and only
synchronize if we are.

This change uncovered a bug in the conformance suite that was making
expecatatation that fall outside of the spec. It has been marked xfail.