[AMP-151] initial _data_host refactoring
The initial work to change the _data_host class to a cache that tracks whether
the home location or gmac buffer has the most recent copy is completed. The
general design is done and it copiles cleanly. However, it causes failure at
runtime because the names of kernels are now different between when it is
compiled for the gpu and compiled for host. This may be a sign of a deepr
issue, but the effectivness of the new design can not be tested until it is
fixed.