[AMP-151] Move changes to _data_host_view
Moved all the changes to _data_host to _data_host_view and add the original
_data_host back in. This lets array keep to the original implimentation without
worrying about the changes made to array_view. Several changes where made to
help conversions between them but there seem to be a few lingering ones.

Also fixed a bug that caused an array and array_view to share the same storage
when an array was asigned from an array_view.