[AMP-151] More refactoring
This commit includes several refactoring changes to pass conformance tests.

1) The type of gmac_buffer_t use T instead of nc_T. The purpose of nc_T was (I
think) to easy conversion, but this has is handled by _data and _data_host_view
now

2) Reinterpret as uses the new method on gmac_buffer_t get_mutable(). Just
using get() was trying to return a const pointer, and then would not convert if
it was being reinterpreted as something non-const. get_mutable() always returns
a mutable pointer.

3) Fix the assertion in refresh.

4) Move the state enum out of the definition of _data_host_view<T> so that the
type of state_ptr is the same for both _data_host_view<T> and
_data_host_view<const T>

5) Make the gmac cache in _data_host_view non-const. Even if we are a const
array view we need to be able to copy into the gmac buffer.

6) Constructors from _data<T> to _data<const T> and
_data_host_view<T> to _data_host_view<const T>