[Refactor] New conformance suite
Adds three new build targets:
amptest: amptest library needed for running tests
conformance: runs the conformance tests listed in the file
amp-conformance/passlists.txt
conformance-all: runs all conformance tests in the Tests directory.

This refactor changes the conformance tests to be based on the original
framework provided in the release of the Microsoft C++amp conformance suite.
Building the test library has been ported to cmake, a large number of fixes
where made to work on linux and clang, and the perl script to run tests was
changed to be configured by cmake and use our tools.

The amptest library and headers had lots of MSVC specific code which was
changed to work on clang. Changed windows '\' directory separators to unix '/'
in the libraries and test cases. Changed wstring to string everywhere.

Changed the run_tests.pl script to work with clamp. It is configured by cmake
to know where to find all the necessary tools. System calls changed to fit
linux instead of MS. The actual compilation is more-or-less a port from the
buildme script in cppamp-sandbox, so changes there will need to be ported back
to this script. It was also extended to take an optional test list and a
version that takes a single testcase from the command line was added to make
iterative testing easier.

Since this replaces the LIT based conformance suite, those cases have been
removed. The target "make test" still exists and will run the unit tests and,
if they have been cloned into the tests directory, the samples. Testing the
conformance suite now requires running the new target "make conformance". To
add a testcase to the conformance testing simply add it to the file
amp-conformance/passlist.txt