[Refactor]Conformance fixes and passlist pruning
Changed a path seperator in an include from '\' to '/' and put the code in
namespace Concurrency::Test in CopyTestFlow.h

Fixed a bug where a configuration file in the same directory as a test was not
found.

Removed the following tests from passlist.txt because they should not compile
but do:
2_Cxx_Lang_Exte/2_4_amp_Rest_Modi/2_4_3_Rest_on_Func_Scop/2_4_3_1_Lit/Negative/Strings.03/test.cpp
2_Cxx_Lang_Exte/2_1_Synt/2_1_3_Type_Spec/Negative/Test.02/test.cpp

Removed the following tests because they incorrectly think they are running on
a platform with shared accelerator memory and fail.
5_Data_Cont/5_1_array/5_1_2_c/CpuSharedMemory/None/test.cpp
5_Data_Cont/5_1_array/5_1_2_c/CpuSharedMemory/Auto/test.cpp

The following tests are output invalid C on CBackend and where not added for any
particular reason. Unsure if they where ever passing on CBackend. They have
been removed for now.
4_Basic_Data_Elmnts/4_1_index/4_1_4_ops/UniOperator/DecrementOperator.01/test.cpp
4_Basic_Data_Elmnts/4_1_index/4_1_4_ops/UniOperator/IncrementOperator.01/test.cpp
4_Basic_Data_Elmnts/4_1_index/4_1_4_ops/Arithmetic/Test.03/test.cpp
4_Basic_Data_Elmnts/4_1_index/4_1_4_ops/Arithmetic/Test.01/test.cpp
4_Basic_Data_Elmnts/4_1_index/4_1_4_ops/Arith_Assign/Test.01/test.cpp
4_Basic_Data_Elmnts/4_1_index/4_1_3_m/Subscript/Test.01/test.cpp
4_Basic_Data_Elmnts/4_1_index/4_1_4_ops/Arith_Assign/Test.03/test.cpp

This test was marked XFAIL in the old system and shouldn't have been added
3_device_Modeling/3_2_Accelerator/3_2_5_members/create_view/test.cpp

This test needs several changes to work in the new system, and may not have
been ported correctly when it was added to the old LIT system. Removed for now.
5_Data_Cont/5_2_a_v/5_2_3_m/WithoutDataSource/Test.03/test.cpp