[AMP-189 Bug] incorrect copy() implementation in amp.h
There are some missing template specializations for copy() in amp.h, and the order of template function declarations would cause clang to fail to instantiate template functions correctly, thus causing some failed cases, such as:

5_Data_Cont/5_3_c_d/Misc/HigherRank/

There are many other cases which would be affected as well.  For the sake of speed of regression test, I only add 1 case in this commit.