[AMP-189 Fix] reorder the order of function templates and add specializations
Fixed the order of various copy() function templates to make sure instantiations work as expected.

Also changed run_test.pl.in to allow more time for a single test case.  The reason for this change is because the test case in this issue would usually take more than 30 seconds to execute.