[AMP-205 Bug] incorrect handling of unsigned long type in GPU build path
2_Cxx_Lang_Exte/2_x_general/Expressions/Pointers/RefToPointer/Normal/Test10/test.cpp

there are 4 configurations for this test case, it would pass when the element type of test data is int, float. It would skip on double as the current hardware doesn't support double. It will FAIL when the element type is "unsigned long".
according to MS, unsigned long is expected to be of 4 bytes. (http://msdn.microsoft.com/en-us/library/s3f49ktz.aspx)
but in SPIR spec, unsigned long is expected to be equivalent with i64.
currently it seems in pre-SPIRify LLVM IR, all unsigned long variables are treated as i32, not i64.