[AMP-237 Fix] change SPIRify so it will handle GEP instructions more carefully
The bug was resulted from the fix to AMP-188.  In that fix it would promote
EVERY GEP instruction.  In this commit I change the logic to make sure we
only promote GEP instructions whose parameters have been promoted.

Also introduced a minor change in the test case to make template instantiation
work.