[AMP-240 Fix] add a helper class to check accelerator associated to an array
add a helper class to array, so it will throw exception if the array is
associated to CPU accelerator.

the way to do it is to implement a class with a custom serializer &
deserializer so when the array is going to be sent to AMP kernels, checking
logic will be invoked.