[AMP-412 Fix] fix CBackend so it won't emit duplicated declarations for pointer-type variables
In the following conformance cases:
2_Cxx_Lang_Exte/2_x_general/Expressions/Pointers/ToShared/Test11.10
2_Cxx_Lang_Exte/2_x_general/Expressions/Pointers/ToShared/Test10

it is observed CBackend will emit duplicated __local variable declarations for pointer-type variables.  For example:

__local int *a;
__local int a;  // cause OpenCL compile error

The bug is caused because CBackend will emit a __local variable whenever it finds a GEP or Store instruction to a variable.  But actually we only need to emit such variables when the operand points to a global value, if the operand points to a local value then we don't need to emit a variable declaration.