[HC] initial implementation of ts_allocator
A new HC interface, ts_allocator is introduced in this commit.

It would take a pointer in the constructor, and provides an alloc() interface
which will allocate the requested size from the pointer.

Right now the implementation only takes the pointer from host memory (global
segment), NOT the desired group segment due to the following limitations:

- HSAIL groupbaseptr instruction does not seem to be working
- obtaining the address of tile_static buffer within kernel would produce
  ptrtoint LLVM IR, but HLC can not lower ptrtoint to HSAIL yet

It is expected the implementation will be moved to use groupbaseptr soon.