[HC] first implementation of accelerator and accelerator_view
This commit is the first implementation for the following classes under the
namespace hc:

- accelerator
- accelerator_view

Right now they are basically the same as their C++AMP counterparts, except for
a new member function which could be used to query the size of group segment
memory.

Test cases are also changed to use these new classes.