[hc test] refactor tests to move completion_future and async_pfe to hc
We want to move non-C++AMP standard APIs from Concurrency namespace to hc.
In this commit, tests for async_parallel_for_each are modified to use hc
namespace instead of Concurrency.  completion_future are also expected to
be in the hc namespace.

Other parameters such as extent, tiled_extent, tiled_index are still in
Concurrency namespace and they will be changed in the future commits.