[hc] refactor how async operations are modelled
Introduce KalmarAsyncOp which forms the abstraction of asynchronous operations
in Kalmar.  Right now only HSA runtime use it but it could be used in OpenCL
as well.  In HSA Kalmar runtime two async operations are modelled as its
subclasses:

- HSADispatch : kernel dispatching logic
- HSABarrier : barrier packet logic

Introduce KalmarEvent which is a POD class.  Each of its instance will hold a
pointer the KalmarAsyncOp object, and the future object returned from its
asynchronous operation.