[Driver] change clamp-link to cope with objcopy 2.25
objcopy 2.25 is different with previous versions that it would complain if an object file contains no sections at all:

https://fossies.org/diffs/binutils/2.24_vs_2.25/binutils/objcopy.c-diff.html

This would produce non-fatal but somewhat messy build log in case binutils 2.25 is used.  In this commit we cope with this new behavior and leave the compiler driver logic as is.