[CMake] separate HSAILasm from llc
up until now CMake scripts assume HSAILasm sit in the same directory with other
tools in HSAIL compiler, but it may not be that way in the real use cases. In
this commit we separate HSAILasm from other tools in HSAIL compiler and
introduce a new CMake option HSAIL_ASSEMBLER_DIR to specify the path to
HSAILasm.