[hc test] Test for "hc_grid_launch" attribute as a string
This test case requires clang to be updated to insert a string
attribute instead of a LLVM first class attribute. "hc_grid_launch"
should be a string attribute (with quotes). Functions that do have
the hc_grid_launch attribute declared in source should also not have
it in IR level.