[hc] Update passes to look for hc_grid_launch as a string attribute
This commit requires clang to be updated to insert hc_grid_launch as
a string attribute. llvm::Function::hasFnAttribute can accept a
StringRef as an argument. llvm::AttrBuilder is used to copy a list
of attributes declared on the cloned function and specifically delete
the "hc_grid_launch" attribute. Then an AttributeSet is created from
AttrBuilder and replaces the attributes in the cloned Function.