[SPIRify] Implementation to cope with loop unroller
In the original implementation, updateListWithUsers() assume that ConstantExpr
 would be immediately used by an Instruction. But with loop unroller in Clang
enabled, it may not be the case. In some C++AMP conformance cases (especially
those in 2_Cxx_Lang_Exte/2_x_general/Statements/ControlFlow/) there are cases
where the following kinds of LLVM IR be produced:

load bitcast (GEP)

GEP is a ConstantExpr
bitcast is another ConstantExpr
load is the Instruction which uses bitcast and GEP

The original implementation can't cope with this scenario and would thus
produce incorrect LLVM IR. In this commit we fix this issue by taking this case
into consideration.

Notice there would be some dead global variables left after the logic. We apply
-globaldce to remove them.