[RemoveSpecialSection] initial commit
This commit introduces a new opt pass: RemoveSepcialSection.

The logic is rather simple, it simply walks through all global variables in a
given module, remove any additional section information attached to them.

The pass is introduced for AMDGPU BE as it is found variables in group segment
can't be in "clamp_opencl_local" section where they originally reside.