[hc test] relax tests related to dynamic group segment allocation
It's discovered that dynamic group segment in the current implementation may
introduce race conditions if multiple kernels in the same module ask for
dynamic group segment.  In this commit we temporarily comment out multiple
kernel dispatches in relevant test cases and mark them as FIXME.

The solution to this issue are still under discussion/debating.