[hc] implement hc::array_view::accelerator_pointer()
Based on the infrastructure built for hc::array::accelerator_pointer(),
implement the same feature for hc::array_view.

There is a known limitation that array_view might not have device memory
associated with it before a kernel launch. So the pointer returned by this
function could be a host pointer. We need a range check function in HSA
runtime to make sure the pointer is a valid device memory pointer.