[pstl] Fix: reduce in lexicographical_compare works incorrectly
There is an dependency for lexicographical comparison.
We need to check each elements sequential.
In the current reduce implementation, the reduction
tree violates this order.
Given tile size = 512
the thread with tile id 0 will first get value
from tile id 256, then tile 128, .... and so on.
The implementation is for performance.
In the specialized reduce version, I reverse the reduction order.
For tile id 0, it will first get value from tile , then tile id 2
and so on.