[WrapperGen] Print 'bool' instead of char for 8-bit integer.
There's an issue with char member in struct as seen here:
https://bitbucket.org/snippets/wukevin/L8nbK
Attempting to compile the above will results in a front end error
with ambient context. Changing it to bool works. Bool still maintains
the same size, so this may be a decent workaround.

Addition:
Remove case for 1-bit integer since there are no corresponding
keywords in C++.