[test] change make test threads from 4 to 2
The latest AMD Radeon OpenCompute Kernel driver has a known issue that its
SDMA engine can only support at most 2 processes using it. So we would run into
trouble if we concurrently run 4 processes which access AMD dGPU with ROCK. In
this commit the number of test threads is reduced from 4 to 2 to temporarily
get around this issue.