[GridLaunch] Fix issue when passing char as an argument to kernel
Due to an earlier commit:
https://bitbucket.org/multicoreware/hcc/pull-requests/164/wrappergen-print-bool-instead-of-char-for/diff
All chars were changed to bool when generating a wrapper file. This
worked because LLVM pads bool to be eight bits when defining a struct.
However, in other situtations, a bool would still be i1. Thus when passing
a char which gets converted to a bool through WrapperGen, LLVM will try
to bitcast it back to char. This will cause opt to crash when running
Promote.

This commit will revert the changes to char so that i1 will become bool
and i8 will become char. A later commit will handle the char issue
within a struct.