Add -O3 when compiling Wrapper code in hc-kernel-assemble
The Wrapper code (*.camp.cpp) is a hidden object that is created by
WrapperGen pass. This code is compiled to a library-like object that
will be linked together with the host code at later stages. Adding a -O3
to this file should increase GridLaunch performance. This will not
affect user passed flags since that happens in hc-kernel-host. The -O3
flag will only be included if hcc is built as Release (including
RelWithDebInfo and MinSizeRel).