[GridLaunch] Use grid_launch_parm_cxx subclass instead of base class when generating the Wrapper
The serializer code is contained in the grid_launch_parm_cxx subclass,
which is only used in the Wrapper. The subclass does not introduce any
new data members and no virtual functions are used. Because the subclass
is passed into the functor, the compiler will be able to utilize the
custom serializer.