Fix handling of invalid/missing tests
If a test list is passed in, and the test is not present, the system
will try to execute that test and return "Invalid or cannot open" .
This will cause the test_results.xml file to truncate at the invalid
test. Handle a non-existent test and ensure that the test is present
before adding it to the list of tests.

Change-Id: Ib9b368aa29f1c8de25b62760d771b05f8054a9a9