[CLANG TOT UPGRADE][OpenCL] remove OpenCL support in HCC
Remove OpenCL support in HCC. As HCC is moving toward ROCm stack as its main
GPU platform, there is no need to support OpenCL anymore. Remove it in this
commit.

Remanents from OpenCL-days may still be found in the codes and will be removed
or refactored in the commits later on.