[Promote] initial version to tackle NAMD build issue
In NAMD, there are complex load/store instructions with 3 levels of
ConstantExpr. Originally at most 2 levels of ConstantExpr were supported. In
this commit, the logic is extended to support 3 levels of ConstantExpr. This
could allow a test case derived from NAMD be successfully compiled.

More additional fixes would be implemented in this branch.