[CMake][Driver] allow ROCm-Device-Libs be configured by CMake
In this commit ROCM_DEVICE_LIB_DIR CMake option can be specified to the
directory where ROCm-Device-Libs is located. It would try find ocml.bc inside
the given directory.

The rationale for this commit is because binary packages for rocm-device-libs
are not made public yet. So required bitcodes may not be present under
/opt/rocm/lib as of now. Users of ToT HCC may have to build ROCm-Device-Libs
themselves and config ToT HCC with ROCM_DEVICE_LIB_DIR option.