SWDEV-123229 HCC RT header needs to trigger re-build
SWDEV-123229 - HCC RT header changes don't trigger a re-build.

When calling "make" to rebuild hcc after modifying the HCC RT headers, the build is not able to detect the change and to rebuild the .cpp files that include the headers being modified.  As a result, developers may be getting an hcc built from stale objects.