Improve Printf implementation to use char arrays on GPU
Allocated a new buffer to store format strings used in printf calls.
Strings created on gpu will be copied into the string buffer on
host coherent memory, and will be printed after kernels. Users can
also write strings directly into arguments of printf call, as shown
in testcase.
TODO: We want to hide the printf implementation from user. Also we
want the printf buffer to be cyclical, so that we can reuse space
that is already printed. There is also a memory access fault that we
run into when the space used is close to buffer size, which needs
further investigation.