Printf - Ordering in Buffers using 1 atomic inst
Previously, the packets were being allocated into the two buffers in different orders with respect to each other. Now using a single atomic compare_exchange_weak, we can maintain ordering in both buffers with respect with each other. Also, string arguments in a single printf, will be allocated consecutively in-order as well. Offsets will be managed in a single ullong, which can be divided into 2 4byte uints. Also, we are now performing a look-ahead to check for the space required to allocation into the string buffer, so we can perform space allocation in printf command chunks.