Improve Link Perf by disabling slp-vectorize
SLP Vectorization is already enabled in Clang frontend ToolChains, so we do not need it for Link time optimizations. This improvement will cut link-time of rocBLAS from 8 mins to 6 mins. Also there is performance increase in run-time of rocblas-test as well.

In detail for compilation time, rocBLAS takes 8 mins for linking, 4 mins for each target. Of the 4 mins, 2 mins spent on opt and 2 mins spent on llc. By removing the slp-vectorize in opt, we will improve opt to run in 1 min.