[RUNTIME] fix potential undefined behavior in mutex locking / unlocking
In the function changed in this commit, originally an RAII lock_guard is
present for a mutex. And in the later part of the function the unlock/lock
member functions of the mutex are directly invoked. Such codes is not ideal.

It was also proven certain user codes would run into soft hang situation.

In this commit we remove the lock_guard and always use lock/unlock member
function calls. The user codes affected above have been proven to run.