Use Binary IR avoids LL non-GlobalValue limit
hc-kernel-assemble.in should use binary IR instead of text IR for the hc-kernel-assemble.in script. Since the build process should not be using .LL files (unless needing to dump IR). This was reveal in recent upstream merge which set limits on non-GlobalValue name lengths which only exists for text LL files.