Unpinned copy engine: use staging buffer by default
The pin-in-place algorithm has race a condition where multiple
threads would try to pin/unpin the same buffer concurrently.
There is no easy way for the hcc runtime to track the pointers
to protect against that and experiments show no performance
benefit with the pin-in-place algorithm.

This patch changes the unpinned copy engine to use the staging
algorithm by default, which is a lot safer.  The staging buffer
size has been adjusted to 4 MB, which seems to be the sweetspot
based on test results.  An environment variable has been added
for adjusting the staging buffer size.

Change-Id: I88ecc9cd254b61d5188e2a8158f10dcbe6ba8436