Flush a fraction of the queue when we are out of space for new kernel (#765)
When the capacity of the asyncOps vector reaches the
MAX_INFLIGHT_COMMANDS_PER_QUEUE threshold, flush only the first
fraction of the vector to overlap the cleanup process (operator
nullification and GC in removeAsyncOp()) with the ongoing GPU
computation.  The fraction size is tunable via the environment
variable HCC_QUEUE_FLUSHING_RATIO.  The default fraction size is half
of the queue.

Change-Id: Ia6a9f4a2fbf07f9944620ee1f275f1f3b324d0d6
Signed-off-by: Sarunya Pumma <sarunya.pumma@amd.com>