Remove deprecated hsa isa compatible call (#896)
* Remove deprecated hsa isa compatible call

Required to use the hsa_agent_iterate_isas for previous deprecated hsa_isa_compatible. Since an agent will allow multiple ISAs from now on, we need to check if code object is compatible with any of them. We leverage passing a struct into the callback function for checking isa handles.

* Use HSA_STATUS_INFO_BREAK for matching ISAs