hcc:Add PATCH number on top of MAJOR/MINOR to dlopen file name
With implementation of adding PATH number to the library file name
absolute file will have libname.so.major.minor.patch as the name.
dlopen need full path if the RPATH or system variables do not have
the info to find the filename. With RPATH updated to libmcamp.so
and hcc binary full path is not necessary.

Change-Id: I972d7c5f32dccdaac9cfefc51d6958723c549c2a
Signed-off-by: Freddy Paul <Freddy.paul@amd.com>