SWDEV-225038: disable relaxed uniform regions in structurizer
The LLVM structurizer includes a local relaxation on the criteria that
define uniform regions, that was introduced in [1] and eventually
enabled by default in [2].

[1] https://reviews.llvm.org/D62198
[2] https://reviews.llvm.org/D63198

But this relaxation is affecting correctness. For example, in
SWDEV-225038, for different combinations of compiler version and app
version, it is possible to either make the app hang or produce
incorrect results.

For now, we disable this relaxation in HCC until we have a complete
resolution in LLVM itself.